+++ src/Controllers/class-packlink-base-controller.php
@@ -7,6 +7,8 @@
 
 namespace Packlink\WooCommerce\Controllers;
 
+use Packlink\BusinessLogic\DTO\BaseDto;
+use Packlink\BusinessLogic\DTO\ValidationError;
 use Packlink\WooCommerce\Components\Utility\Shop_Helper;
 
 /**
@@ -21,6 +23,17 @@ class Packlink_Base_Controller {
 	 * @var bool
 	 */
 	protected $is_internal = true;
+	/**
+	 * Translation messages for fields that are being validated.
+	 *
+	 * @var array
+	 */
+	private $validation_messages = array(
+		'email'       => 'Field must be valid email.',
+		'phone'       => 'Field must be valid phone number.',
+		'weight'      => 'Weight must be a positive decimal number.',
+		'postal_code' => 'Postal code is not correct.',
+	);
 
 	/**
 	 * Processes request. Reads 'action' parameter and calls action method if provided.
@@ -57,6 +70,56 @@ class Packlink_Base_Controller {
 
 			exit();
 		}
+	}
+
+	/**
+	 * Converts DTOs to array and returns a JSON response.
+	 *
+	 * @param BaseDto[] $entities
+	 */
+	protected function return_dto_entities_response( array $entities ) {
+		$response = array();
+
+		foreach ( $entities as $entity ) {
+			$response[] = $entity->toArray();
+		}
+
+		$this->return_json( $response );
+	}
+
+	/**
+	 * Returns 400 response with validation errors.
+	 *
+	 * @param ValidationError[] $errors
+	 */
+	protected function return_validation_errors_response( array $errors ) {
+		$response = array();
+
+		foreach ( $errors as $error ) {
+			$response[$error->field] = $this->get_validation_error_message( $error->code, $error->field );
+		}
+
+		$this->return_json( $response, 400 );
+	}
+
+	/**
+	 * Returns a validation message for validation error.
+	 *
+	 * @param string $code
+	 * @param string $field
+	 *
+	 * @return string
+	 */
+	private function get_validation_error_message( $code, $field ) {
+		if ( ValidationError::ERROR_REQUIRED_FIELD === $code ) {
+			return __( 'Field is required.', 'packlink-pro-shipping' );
+		}
+
+		if ( in_array( $code, array( 'width', 'length', 'height' ) ) ) {
+			return __( 'Field must be valid whole number.', 'packlink-pro-shipping' );
+		}
+
+		return __( $this->validation_messages[ $field ], 'packlink-pro-shipping' );
 	}
 
 	/**